+++ 0x06-pointers_arrays_strings/main.h
@@ -7,7 +7,7 @@
 
 char *_strcat(char *dest, char *src);
 int _putchar(char c);
-char *_strncat(char *dest, char *src, size_t n);
+char *_strncat(char *dest, char *src, int n);
 
 
 #endif
